requirements.txt
adjustText==0.7.3
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
backports.zoneinfo==0.2.1
bioinfokit==2.0.8
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
channels==2.4.0
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
cycler==0.11.0
daphne==2.5.0
Deprecated==1.2.13
django-redis==5.2.0
dpd-components==0.1.0
et-xmlfile==1.1.0
ez-setup==0.9
fonttools==4.33.3
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
incremental==21.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
lazy_loader==0.1rc2
littleutils==0.2.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-venn==0.11.7
mpmath==1.1.0
msgpack==1.0.4
openpyxl==3.0.10
outdated==0.2.1
packaging==21.3
pandas==1.1.5
pandas-flavor==0.3.0
patsy==0.5.2
Pillow==9.2.0
plotly==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
service-identity==21.1.0
six==1.16.0
sklearn 
sqlparse==0.4.2
statsmodels==0.13.2
tabulate==0.8.10
tenacity==8.0.1
textwrap3==0.9.2
threadpoolctl==3.1.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.0.3
wrapt==1.14.1
xarray==2022.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
zope.interface==5.4.0


